Rename folder for case-sensitivity
